requirements.txt
streamlit==1.44.0
plotly==6.0.1
pandas==2.2.3
numpy==2.2.1
folium==0.19.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
haversine==2.9.0
streamlit-folium==0.24.0
Pillow==11.1.0
inflection==0.5.1